python-bg-remover/requirements.txt
flask
flask-cors
rembg
pillow
requests
